+++ src/main/java/br/com/prevodonto/service/impl/ClienteServiceImpl.java
@@ -27,7 +27,7 @@ public class ClienteServiceImpl implements ClienteService {
 	}
 
 	@Override
-	public void deletar(Integer id) throws NotFoundException {
+	public void deletar(Long id) throws NotFoundException {
 		Cliente clienteEncontrado = this.repository.findById(id)
 				.orElseThrow(() -> new NotFoundException("Cliente não encontrado"));
 		this.repository.delete(clienteEncontrado);
@@ -38,7 +38,7 @@ public class ClienteServiceImpl implements ClienteService {
 		return this.repository.findByCpf(cpf).orElseThrow(() -> new NotFoundException("Cpf não encontrado"));
 	}
 
-	public Cliente find(Integer id) {
+	public Cliente find(Long id) {
 		Optional<Cliente> obj = repository.findById(id);
 		return obj.orElseThrow(()-> new ObjectNotFoundException("Objeto não encontrado: Id: " +id + Cliente.class.getName()));
 	}
